YAY-114 refactor: transfer request from setting.service to couple .sevice
