+++ src/features/settings/api/index.ts
@@ -1,0 +1,4 @@
+export * from './devices.service'
+export * from './payments.service'
+export * from './profile.service'
+export * from './settings.types'
